requirement.txt
absl-py==0.2.0
astor==0.6.2
atari-py==0.1.1
attrs==18.1.0
awscli==1.15.19
backcall==0.1.0
bleach==2.1.3
boto3==1.7.19
botocore==1.10.19
Box2D==2.3.1
cached-property==1.4.2
certifi==2018.4.16
cffi==1.11.5
chainer==1.18.0
chardet==3.0.4
click==6.6
cloudpickle==0.5.2
colorama==0.3.7
coverage==4.5.1
cycler==0.10.0
Cython==0.28.2
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
filelock==3.0.4
Flask==1.0.2
future==0.16.0
gast==0.2.0
grpcio==1.11.0
gym==0.7.4
h5py==2.7.1
html5lib==0.9999999
hyperopt==0.1
idna==2.6
imageio==2.3.0
ipdb==0.11
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==5.1.5
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.10.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==1.2.1
Lasagne==0.2.dev1
line-profiler==2.1.2
llvmlite==0.22.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.8.3
mock==2.0.0
more-itertools==4.1.0
mpi4py==2.0.0
msgpack-python==0.5.6
mujoco-py==0.5.7
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nibabel==2.1.0
nine==1.0.0
nose==1.3.7
nose2==0.7.4
notebook==5.4.1
numba==0.37.0
numpy==1.12.0
numpy-stl==2.2.0
olefile==0.45.1
pandas==0.21.1
pandocfilters==1.4.2
parso==0.2.0
path.py==11.0.1
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==3.2.0
plotly==1.9.6
pluggy==0.6.0
polling==0.3.0
prettytensor==0.6.2
progressbar2==3.37.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pycparser==2.18
pygame==1.9.2a0
pyglet==1.3.2
Pygments==2.2.0
pylru==1.0.9
pymongo==3.6.1
PyOpenGL==3.1.0
pyparsing==2.2.0
PyPrind==2.11.2
pytest==3.5.1
python-dateutil==2.7.2
python-utils==2.3.0
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
rllab==0.1.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==0.19.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==0.9.0.dev1
torch==0.1.9+aaf41c6
torchvision==0.1.6
tornado==5.0.2
tqdm==4.23.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==1.2.3
